# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation des APIs publiques.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/{id}": {
            "get": {
                "summary": "Afficher un sweatshirt sp\u00e9cifique.",
                "operationId": "get_admin_sweatshirt_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "summary": "Supprimer un sweatshirt.",
                "operationId": "post_admin_sweatshirt_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/{id}/edit": {
            "get": {
                "summary": "Modifier un sweatshirt.",
                "operationId": "get_admin_sweatshirt_edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "summary": "Modifier un sweatshirt.",
                "operationId": "post_admin_sweatshirt_edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}